Chatbot/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
httpx==0.23.0
authlib==1.0.0
github==3.1.0
langchain-groq==0.1.0
pyjwt==2.8.0
streamlit==1.18.0
